refactor cli helpers
