+++ website/src/plugins/pluginPlayground/cli/helpers/getDependencies.ts
