+++ website/src/plugins/pluginPlayground/cli/helpers/getMdxFromMarkdownString.ts
@@ -1,10 +1,9 @@
-// these packages are avail from rspress
 import { mdxFromMarkdown } from 'mdast-util-mdx'
 import { mdxjs } from 'micromark-extension-mdxjs'
 import { fromMarkdown } from 'mdast-util-from-markdown'
 
-export const getMdxFromMarkdown = (doc: string) =>
-  fromMarkdown(doc, {
+export const getMdxFromMarkdownString = (mdString: string) =>
+  fromMarkdown(mdString, {
     extensions: [mdxjs()],
     mdastExtensions: [mdxFromMarkdown()],
   })
