+++ website/src/plugins/pluginPlayground/cli/helpers/getMdxJsxAttribute.ts
@@ -1,0 +1,9 @@
+import { MdxJsxFlowElement } from 'mdast-util-mdx'
+
+export const getMdxJsxAttribute = (node: MdxJsxFlowElement, attrName: string) => {
+  const attribute = node.attributes.find((attr) => {
+    return attr.type === 'mdxJsxAttribute' && attr.name === attrName
+  })
+
+  return attribute?.value
+}
