+++ website/src/plugins/pluginPlayground/cli/helpers/transformAssetPaths.ts
