--- website/src/plugins/pluginPlayground/cli/utils/getNodeAttribute.ts
@@ -1,4 +1,0 @@
-export const getNodeAttribute = (node: any, attrName: string): string | undefined => {
-  return node.attributes.find((attr: { name: string; value: string }) => attr.name === attrName)
-    ?.value
-}
